+++ modules/transit-gateway-apis/src/main/java/com/ibm/cloud/networking/transit_gateway_apis/v1/model/TransitGatewayConnectionCust.java
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright IBM Corp. 2020.
+ * (C) Copyright IBM Corp. 2021.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -23,18 +23,22 @@ import com.ibm.cloud.sdk.core.service.model.GenericModel;
 public class TransitGatewayConnectionCust extends GenericModel {
 
   /**
-   * Defines what type of network is connected via this connection.
+   * Defines what type of network is connected via this connection. The list of enumerated values for this property may
+   * expand in the future. Code and processes using this field must tolerate unexpected values.
    */
   public interface NetworkType {
     /** vpc. */
     String VPC = "vpc";
     /** classic. */
     String CLASSIC = "classic";
+    /** gre_tunnel. */
+    String GRE_TUNNEL = "gre_tunnel";
   }
 
   /**
-   * Only visible for cross account connections, this field represents the status of the request to connect the given
-   * network between accounts.
+   * Only visible for cross account connections, this field represents the status of a connection request between IBM
+   * Cloud accounts. The list of enumerated values for this property may expand in the future. Code and processes using
+   * this field must tolerate unexpected values.
    */
   public interface RequestStatus {
     /** pending. */
@@ -50,7 +54,8 @@ public class TransitGatewayConnectionCust extends GenericModel {
   }
 
   /**
-   * What is the current configuration state of this connection.
+   * Connection's current configuration state. The list of enumerated values for this property may expand in the future.
+   * Code and processes using this field must tolerate unexpected values.
    */
   public interface Status {
     /** attached. */
@@ -72,22 +77,37 @@ public class TransitGatewayConnectionCust extends GenericModel {
   protected String networkId;
   @SerializedName("network_type")
   protected String networkType;
-  @SerializedName("network_account_id")
-  protected String networkAccountId;
   protected String id;
+  @SerializedName("base_connection_id")
+  protected String baseConnectionId;
   @SerializedName("created_at")
   protected Date createdAt;
+  @SerializedName("local_bgp_asn")
+  protected Long localBgpAsn;
+  @SerializedName("local_gateway_ip")
+  protected String localGatewayIp;
+  @SerializedName("local_tunnel_ip")
+  protected String localTunnelIp;
+  protected Long mtu;
+  @SerializedName("network_account_id")
+  protected String networkAccountId;
+  @SerializedName("remote_bgp_asn")
+  protected Long remoteBgpAsn;
+  @SerializedName("remote_gateway_ip")
+  protected String remoteGatewayIp;
+  @SerializedName("remote_tunnel_ip")
+  protected String remoteTunnelIp;
   @SerializedName("request_status")
   protected String requestStatus;
   protected String status;
   @SerializedName("updated_at")
   protected Date updatedAt;
+  protected TransitGatewayConnectionCustZone zone;
 
   /**
    * Gets the name.
    *
-   * The user-defined name for this transit gateway. If unspecified, the name will be the network name (the name of the
-   * VPC in the case of network type 'vpc', and the word Classic, in the case of network type 'classic').
+   * The user-defined name for this transit gateway connection.
    *
    * @return the name
    */
@@ -99,8 +119,7 @@ public class TransitGatewayConnectionCust extends GenericModel {
    * Gets the networkId.
    *
    * The ID of the network being connected via this connection. This field is required for some types, such as 'vpc'.
-   * For network type 'vpc' this is the CRN of the VPC to be connected. This field is required to be unspecified for
-   * network type 'classic'.
+   * For network type 'vpc' this is the CRN of the VPC to be connected.
    *
    * @return the networkId
    */
@@ -111,7 +130,8 @@ public class TransitGatewayConnectionCust extends GenericModel {
   /**
    * Gets the networkType.
    *
-   * Defines what type of network is connected via this connection.
+   * Defines what type of network is connected via this connection. The list of enumerated values for this property may
+   * expand in the future. Code and processes using this field must tolerate unexpected values.
    *
    * @return the networkType
    */
@@ -120,26 +140,28 @@ public class TransitGatewayConnectionCust extends GenericModel {
   }
 
   /**
-   * Gets the networkAccountId.
-   *
-   * The ID of the account which owns the network that is being connected. Generally only used if the network is in a
-   * different account than the gateway.
-   *
-   * @return the networkAccountId
-   */
-  public String getNetworkAccountId() {
-    return networkAccountId;
-  }
-
-  /**
    * Gets the id.
    *
-   * The unique identifier for this Transit Gateway Connection to Network (vpc/classic).
+   * The unique identifier for this Transit Gateway Connection.
    *
    * @return the id
    */
   public String getId() {
     return id;
+  }
+
+  /**
+   * Gets the baseConnectionId.
+   *
+   * network_type 'gre_tunnel' connections use 'base_connection_id' to specify the ID of a network_type 'classic'
+   * connection the tunnel is configured over. The specified connection must reside in the same transit gateway and be
+   * in an active state. The 'classic' connection cannot be deleted until any 'gre_tunnel' connections using it are
+   * deleted. This field only applies to and is required for network type 'gre_tunnel' connections.
+   *
+   * @return the baseConnectionId
+   */
+  public String getBaseConnectionId() {
+    return baseConnectionId;
   }
 
   /**
@@ -154,10 +176,100 @@ public class TransitGatewayConnectionCust extends GenericModel {
   }
 
   /**
+   * Gets the localBgpAsn.
+   *
+   * Local network BGP ASN.  This field only applies to network type 'gre_tunnel' connections.
+   *
+   * @return the localBgpAsn
+   */
+  public Long getLocalBgpAsn() {
+    return localBgpAsn;
+  }
+
+  /**
+   * Gets the localGatewayIp.
+   *
+   * Local gateway IP address.  This field only applies to network type 'gre_tunnel' connections.
+   *
+   * @return the localGatewayIp
+   */
+  public String getLocalGatewayIp() {
+    return localGatewayIp;
+  }
+
+  /**
+   * Gets the localTunnelIp.
+   *
+   * Local tunnel IP address.  This field only applies to network type 'gre_tunnel' connections.
+   *
+   * @return the localTunnelIp
+   */
+  public String getLocalTunnelIp() {
+    return localTunnelIp;
+  }
+
+  /**
+   * Gets the mtu.
+   *
+   * GRE tunnel MTU.  This field only applies to network type 'gre_tunnel' connections.
+   *
+   * @return the mtu
+   */
+  public Long getMtu() {
+    return mtu;
+  }
+
+  /**
+   * Gets the networkAccountId.
+   *
+   * The ID of the account which owns the connected network. Generally only used if the network is in a different IBM
+   * Cloud account than the gateway.
+   *
+   * @return the networkAccountId
+   */
+  public String getNetworkAccountId() {
+    return networkAccountId;
+  }
+
+  /**
+   * Gets the remoteBgpAsn.
+   *
+   * Remote network BGP ASN.  This field only applies to network type 'gre_tunnel' connections.
+   *
+   * @return the remoteBgpAsn
+   */
+  public Long getRemoteBgpAsn() {
+    return remoteBgpAsn;
+  }
+
+  /**
+   * Gets the remoteGatewayIp.
+   *
+   * Remote gateway IP address.  This field only applies to network type 'gre_tunnel' connections.
+   *
+   * @return the remoteGatewayIp
+   */
+  public String getRemoteGatewayIp() {
+    return remoteGatewayIp;
+  }
+
+  /**
+   * Gets the remoteTunnelIp.
+   *
+   * Remote tunnel IP address.  This field only applies to network type 'gre_tunnel' connections.
+   *
+   * @return the remoteTunnelIp
+   */
+  public String getRemoteTunnelIp() {
+    return remoteTunnelIp;
+  }
+
+  /**
    * Gets the requestStatus.
    *
-   * Only visible for cross account connections, this field represents the status of the request to connect the given
-   * network between accounts.
+   * Only visible for cross account connections, this field represents the status of a connection request between IBM
+   * Cloud accounts. The list of enumerated values for this property may expand in the future. Code and processes using
+   * this field must tolerate unexpected values.
    *
    * @return the requestStatus
    */
@@ -168,7 +280,8 @@ public class TransitGatewayConnectionCust extends GenericModel {
   /**
    * Gets the status.
    *
-   * What is the current configuration state of this connection.
+   * Connection's current configuration state. The list of enumerated values for this property may expand in the future.
+   * Code and processes using this field must tolerate unexpected values.
    *
    * @return the status
    */
@@ -185,6 +298,17 @@ public class TransitGatewayConnectionCust extends GenericModel {
    */
   public Date getUpdatedAt() {
     return updatedAt;
+  }
+
+  /**
+   * Gets the zone.
+   *
+   * Location of GRE tunnel.  This field only applies to network type 'gre_tunnel' connections.
+   *
+   * @return the zone
+   */
+  public TransitGatewayConnectionCustZone getZone() {
+    return zone;
   }
 }