+++ modules/transit-gateway-apis/src/test/java/com/ibm/cloud/networking/transit_gateway_apis/v1/model/TransitGatewayConnectionCustTest.java
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright IBM Corp. 2020.
+ * (C) Copyright IBM Corp. 2021.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -14,6 +14,7 @@
 package com.ibm.cloud.networking.transit_gateway_apis.v1.model;
 
 import com.ibm.cloud.networking.transit_gateway_apis.v1.model.TransitGatewayConnectionCust;
+import com.ibm.cloud.networking.transit_gateway_apis.v1.model.TransitGatewayConnectionCustZone;
 import com.ibm.cloud.networking.transit_gateway_apis.v1.utils.TestUtilities;
 
 import com.ibm.cloud.sdk.core.service.model.FileWithMetadata;
@@ -38,11 +39,20 @@ public class TransitGatewayConnectionCustTest {
     assertNull(transitGatewayConnectionCustModel.getName());
     assertNull(transitGatewayConnectionCustModel.getNetworkId());
     assertNull(transitGatewayConnectionCustModel.getNetworkType());
-    assertNull(transitGatewayConnectionCustModel.getNetworkAccountId());
     assertNull(transitGatewayConnectionCustModel.getId());
+    assertNull(transitGatewayConnectionCustModel.getBaseConnectionId());
     assertNull(transitGatewayConnectionCustModel.getCreatedAt());
+    assertNull(transitGatewayConnectionCustModel.getLocalBgpAsn());
+    assertNull(transitGatewayConnectionCustModel.getLocalGatewayIp());
+    assertNull(transitGatewayConnectionCustModel.getLocalTunnelIp());
+    assertNull(transitGatewayConnectionCustModel.getMtu());
+    assertNull(transitGatewayConnectionCustModel.getNetworkAccountId());
+    assertNull(transitGatewayConnectionCustModel.getRemoteBgpAsn());
+    assertNull(transitGatewayConnectionCustModel.getRemoteGatewayIp());
+    assertNull(transitGatewayConnectionCustModel.getRemoteTunnelIp());
     assertNull(transitGatewayConnectionCustModel.getRequestStatus());
     assertNull(transitGatewayConnectionCustModel.getStatus());
     assertNull(transitGatewayConnectionCustModel.getUpdatedAt());
+    assertNull(transitGatewayConnectionCustModel.getZone());
   }
 }